Test_OpenAI_API/requirements.txt
fastapi
uvicorn
jinja2
openai>=1.0.0
python-dotenv